feat: re-derive the mixnet projection from the typed status snapshot
The seed slice deliberately excluded the old mixnet projection: it
rendered a dead tri-state and re-minted the retired off token, so it
could only be re-derived, never ported. This slice derives it from
zingolib typed MixnetStatus. The status line absorbs zingo-cli
render_status wording verbatim with its three pinned tests; the
disclaimer composition absorbs render_status_with_disclaimer; the JSON
view is a parse-and-render of the golden-pinned serde wire, so no
second JSON vocabulary can drift into existence; and a snapshot ext
trait reads the session status channel whole. The CLI nym status arm
delegates to the projection and its local renderers retire.
